+++ app/vodloader/api.py
@@ -1,7 +1,7 @@
 from quart import Blueprint, Quart, request, current_app
-from .vodloader import *
-from .channel import ChannelDoesNotExist
 from os import environ
+from vodloader.models import TwitchClient
+from twitchAPI.helper import first
 
 
 api = Blueprint('api', __name__)
@@ -10,35 +10,30 @@ api = Blueprint('api', __name__)
 @api.route("/channel", methods=['POST'])
 async def add_channel():
 
-    data = await request.get_json()
-
-    if 'secret' not in request.headers or request.headers['secret'] != current_app.secret_key:
-        return 'Get outta here ya bum', 403
-
-    if 'channel' not in data :
-        return 'The "channel" field is required', 400
-    else:
-        channel = data['channel']
-
-    if 'quality' in data:
-        quality = data['quality']
-    else:
-        quality = 'best'
-    
-    vodloader: VODLoader = current_app.config['vodloader']
     try:
-        await vodloader.add_channel(channel, quality)
-    except ChannelAlreadyAdded as e:
-        return {
-            'result': 'failure',
-            'reason': f'The channel "{channel}" has already been added to VODLoader'
-        }
-    except ChannelDoesNotExist as e:
-        return {
-            'result': 'failure',
-            'reason': f'The channel "{channel}" does not exist on Twitch'
-        }
-    return {'result': 'success'}
+        data = await request.get_json()
+
+        if 'secret' not in request.headers or request.headers['secret'] != current_app.secret_key:
+            return 403
+
+        if 'channel' not in data :
+            return 'The "channel" field is required', 400
+        else:
+            channel_name = data['channel']
+
+        if 'quality' in data:
+            quality = data['quality']
+        else:
+            quality = 'best'
+
+        twitch = await TwitchClient.get_twitch()
+        channel = await first(twitch.get_users(logins=[channel_name]))
+
+    except Exception as e:
+        return 500
+    
+    return 200
+
 
 @api.route("/channel/<channel>", methods=['DELETE'])
 async def delete_channel(channel: str):
@@ -57,6 +52,7 @@ async def delete_channel(channel: str):
         }
     return {'result': 'success'}
 
+
 @api.route("/channels", methods=['GET'])
 async def get_channels():
     
@@ -71,9 +67,9 @@ async def get_channels():
     }
     return channels
 
-def create_api(vodloader: VODLoader) -> Quart:
+
+def create_api() -> Quart:
     app = Quart(__name__)
-    app.config['vodloader'] = vodloader
     if 'API_SECRET_KEY' not in environ or not environ['API_SECRET_KEY']:
         raise RuntimeError('API_SECRET_KEY must be specified either as an environment variable or in the .env file')
     app.secret_key = environ['API_SECRET_KEY']
